Strongly typed transactions in handlers
During transaction processing we retrieve the type of msg and based on
that we get the applicable handler from the router. This means that the
router should only receive messages that it knows how to handle. Instead
of using Get interfaces, we should cast to the actual type of
transaction and then access the transaction details that way.
It's okay to panic here, because if the DummyHandler receives a message
that it cannot cast to the expected type it means something is wrong
with the router. Instead of retrieving an arbitrary key by chance we
should panic.